# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: VOUCHERBASE
  version: 1.0.0
  description: Parrish Lyon. Property planning only; not PHA approval or household subsidy determination.
servers:
  - url: http://localhost:8080/api/v1
security:
  - bearerKey: []
paths:
  /me:
    get:
      responses: {'200': {description: Authenticated tenant, role, version and demo state}, '401': {description: Invalid key}}
  /properties:
    get:
      responses: {'200': {description: Up to 1000 tenant properties with latest saved result}}
    post:
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/Property'}}}
      responses: {'201': {description: Immutable property specification created}, '422': {description: Invalid fields}}
  /analyses:
    get:
      responses: {'200': {description: Up to 200 saved analysis references}}
    post:
      parameters: [{$ref: '#/components/parameters/Idempotency'}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/Analysis'}}}
      responses:
        '201': {description: Saved result with source rows, hashes and engine request}
        '200': {description: Identical idempotent request returned existing result}
        '409': {description: Key reused with different input}
        '422': {description: Unsupported calculation or invalid assumptions}
  /analyses/{id}:
    parameters: [{$ref: '#/components/parameters/ID'}]
    get:
      responses: {'200': {description: Saved input and result}, '404': {description: Missing or inaccessible}}
  /analyses/{id}/replay:
    parameters: [{$ref: '#/components/parameters/ID'}]
    post:
      responses: {'200': {description: Replay hash comparison; no fresh source acquisition}}
  /jobs:
    get:
      responses: {'200': {description: Up to 100 tenant job summaries}}
    post:
      parameters: [{$ref: '#/components/parameters/Idempotency'}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [items]
              properties: {items: {type: array, minItems: 1, maxItems: 1000, items: {$ref: '#/components/schemas/Analysis'}}}
      responses: {'202': {description: Durable batch accepted}, '409': {description: Idempotency conflict}}
  /jobs/{id}:
    parameters: [{$ref: '#/components/parameters/ID'}]
    get:
      responses: {'200': {description: Summary and per-item state, attempts, result and error}}
  /catalog:
    get:
      responses: {'200': {description: Authorities, source metadata, changes and coverage limits}}
  /sources/{id}:
    parameters: [{$ref: '#/components/parameters/ID'}]
    get:
      responses: {'200': {description: Stored snapshot; does not fetch arbitrary URLs}}
  /audit:
    get:
      responses: {'200': {description: Up to 200 tenant audit events}}
  /geocode:
    post:
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [address, provider]
              properties: {address: {type: string, maxLength: 500}, provider: {enum: [census, fixture]}}
      responses: {'200': {description: Labeled matches or explicit disabled/no-match state}}
  /authorities:
    get:
      parameters:
        - {in: query, name: latitude, required: true, schema: {type: number}}
        - {in: query, name: longitude, required: true, schema: {type: number}}
        - {in: query, name: state, required: true, schema: {type: string}}
      responses: {'200': {description: Candidate authorities; confirmation always required}}
  /imports/properties:
    post:
      requestBody:
        required: true
        content: {text/csv: {schema: {type: string, description: See downloadable property-template.csv. Maximum 2 MB and 1000 rows; atomic import.}}}
      responses: {'201': {description: Imported count}, '422': {description: Rejected with no partial import}}
  /comparables:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rows]
              properties:
                rows:
                  type: array
                  maxItems: 1000
                  minItems: 1
                  items:
                    allOf:
                      - {$ref: '#/components/schemas/Property'}
                      - type: object
                        required: [external_id, source_url, kind, observed_on, rent_cents, property_ref]
                        properties:
                          external_id: {type: string}
                          source_url: {type: string, format: uri}
                          kind: {enum: [asking, executed, synthetic]}
                          observed_on: {type: string, format: date}
                          rent_cents: {type: integer, minimum: 1, maximum: 1000000000}
                          property_ref: {type: string}
      responses: {'201': {description: Imported observations}, '409': {description: Duplicate external ID}}
components:
  securitySchemes:
    bearerKey: {type: http, scheme: bearer, description: High-entropy provisioned landlord key}
  parameters:
    ID: {in: path, name: id, required: true, schema: {type: string, format: uuid}}
    Idempotency: {in: header, name: Idempotency-Key, required: true, schema: {type: string, minLength: 1, maxLength: 150}}
  schemas:
    Property:
      type: object
      required: [address, city, state, zip, bedrooms, bathrooms, sqft, structure, condition, geocode_status, is_demo]
      properties:
        id: {type: string, format: uuid, readOnly: true}
        address: {type: string, minLength: 3, maxLength: 500}
        city: {type: string}
        state: {type: string, pattern: '^[A-Z]{2}$'}
        zip: {type: string, pattern: '^\d{5}$'}
        bedrooms: {type: integer, minimum: 0, maximum: 6}
        bathrooms: {type: number, exclusiveMinimum: 0, maximum: 20}
        sqft: {type: integer, minimum: 100, maximum: 30000}
        structure: {enum: [detached, attached, high_rise]}
        condition: {enum: [poor, fair, average, good, renovated]}
        amenities: {type: array, maxItems: 20, items: {type: string}}
        latitude: {type: [number, 'null']}
        longitude: {type: [number, 'null']}
        geocode_status: {enum: [unverified, user_confirmed, census_interpolated, synthetic]}
        is_demo: {type: boolean}
    Analysis:
      type: object
      additionalProperties: false
      required: [property_id, effective_on, comparable_kind, utilities]
      properties:
        property_id: {type: string, format: uuid}
        effective_on: {type: string, format: date}
        knowledge_at: {type: string, description: RFC3339 timestamp or empty for receipt time}
        authority_id: {type: string, maxLength: 32}
        authority_confirmed: {type: boolean}
        voucher_bedrooms: {type: [integer, 'null'], minimum: 0, maximum: 6}
        proposed_rent_cents: {type: [integer, 'null'], minimum: 0, maximum: 1000000000000}
        comparable_kind: {enum: [asking, executed, synthetic]}
        benchmark_entity: {type: string}
        utilities:
          type: object
          additionalProperties: false
          properties:
            complete: {type: boolean, description: True means responsibility selections are complete; empty complete list explicitly means owner pays all}
            items: {type: array, items: {type: string, description: Utility codes from the accepted authority schedule; conflicting fuel/service selections rejected}}
        underwriting: {type: [object, 'null'], description: All fields required when present. See Rust Underwriting type and contracts/example-analysis.json. Money is integer cents; ratios are basis points.}
        scenarios: {type: array, maxItems: 32, items: {type: object, description: name plus optional rent_cents, vacancy_bps, renovation_cents}}